Split into separate files, quick fixes
