+++ package.json
@@ -45,7 +45,7 @@
   },
   "dependencies": {
     "@chainsafe/bls": "^0.1.7",
-    "@chainsafe/eth2.0-types": "^0.1.1",
+    "secure-random": "^1.1.2",
     "uuid": "^3.3.3"
   }
 }
