+++ src/keystore.ts
@@ -1,62 +1,10 @@
 import uuid from "uuid";
-import { scrypt, ScryptParams, PBKDF2Params, PBKDF2, DefaultPBKDF2Params, AES_128_CTR, SHA256, DefaultScryptParams } from "./utils/crypto";
-import {bytes} from "@chainsafe/eth2.0-types";
-import { randomBytes } from "crypto";
-import {PublicKey, PrivateKey} from "@chainsafe/bls";
-
-export interface IKeystoreModule{
-  function: string;
-  params: ScryptParams | PBKDF2Params | any;
-  message: bytes;
-}
-
-export class KeystoreModule implements IKeystoreModule {
-  public function = "";
-  public params: ScryptParams | PBKDF2Params | any = {};
-  public message: bytes = new Buffer("");
-
-  public static fromJson(json: IKeystoreModule): KeystoreModule {
-    const keystoreModule = new KeystoreModule();
-    keystoreModule.function = json["function"];
-    keystoreModule.params = json["params"];
-    keystoreModule.message = json["message"];
-
-    return keystoreModule;
-  }
-}
-
-export interface IKeystoreCrypto{
-  kdf: IKeystoreModule;
-  checksum: IKeystoreModule;
-  cipher: IKeystoreModule;
-}
-
-export class KeystoreCrypto implements IKeystoreCrypto {
-
-  public kdf: IKeystoreModule = new KeystoreModule();
-  public checksum: IKeystoreModule = new KeystoreModule();
-  public cipher: IKeystoreModule = new KeystoreModule();
-
-  public static fromJson(json: IKeystoreCrypto): KeystoreCrypto {
-    const keystoreCrypto = new KeystoreCrypto();
-
-    keystoreCrypto.kdf = KeystoreModule.fromJson(json["kdf"]);
-    keystoreCrypto.checksum = KeystoreModule.fromJson(json["checksum"]);
-    keystoreCrypto.cipher = KeystoreModule.fromJson(json["cipher"]);
-
-    return keystoreCrypto;
-  }
-}
-
-export interface IKeystore{
-  crypto: IKeystoreCrypto;
-  pubkey: string;
-  path: string;
-  uuid: string;
-  version: number;
-
-  decrypt(password: string): Buffer;
-}
+import { scrypt, PBKDF2, DefaultPBKDF2Params, AES_128_CTR, SHA256, DefaultScryptParams } from "./utils/crypto";
+import { PublicKey, PrivateKey } from "@chainsafe/bls";
+// @ts-ignore
+import secureRandom from "secure-random";
+import { IKeystoreCrypto, ScryptParams, PBKDF2Params, bytes, IKeystore } from ".";
+import { KeystoreCrypto } from "./keystore-crypto";
 
 export class Keystore implements IKeystore {
   public crypto: IKeystoreCrypto = new KeystoreCrypto();
@@ -65,11 +13,14 @@ export class Keystore implements IKeystore {
   public uuid: string = uuid.v4();
   public version = 4;
 
-  private kdf(password: string, args: object): Buffer{
-    if(this.crypto.kdf.function === "scrypt"){
-      return scrypt(password, args as ScryptParams);
-    }else{
-      return PBKDF2(password, args as PBKDF2Params);
+  private kdf(password: string, args: ScryptParams | PBKDF2Params): Buffer {
+    switch(this.crypto.kdf.function){
+      case "scrypt":
+        return scrypt(password, args as ScryptParams);
+      case "pbkdf2":
+        return PBKDF2(password, args as PBKDF2Params);
+      default:
+        throw new Error("Unsupported crypto function");
     }
   }
 
@@ -84,21 +35,21 @@ export class Keystore implements IKeystore {
     return keystore;
   }
 
-  public static encrypt(secret: bytes, password: string, path = "", kdfSalt: bytes = randomBytes(32), aesIv: bytes = randomBytes(16)): IKeystore {
+  public static encrypt(secret: bytes, password: string, path = "", kdfSalt: bytes = secureRandom(32, { type: "Buffer" }), aesIv: bytes = secureRandom(16, { type: "Buffer" })): IKeystore {
     const keystore = new this();
 
     keystore.crypto.kdf.params.salt = kdfSalt;
     const decryptionKey: bytes = keystore.kdf(password, keystore.crypto.kdf.params);
     keystore.crypto.cipher.params.iv = aesIv;
     const cipher = AES_128_CTR(decryptionKey.slice(0, 16), keystore.crypto.cipher.params.iv);
-        
+
     let encryptedSecret = cipher.update(secret);
     encryptedSecret = Buffer.concat([encryptedSecret, cipher.final()]);
-        
+
     keystore.crypto.cipher.message = Buffer.from(encryptedSecret);
     keystore.crypto.checksum.message = SHA256(Buffer.concat([decryptionKey.slice(16, 32), keystore.crypto.cipher.message]));
-        
-    keystore.pubkey = PublicKey.fromPrivateKey(PrivateKey.fromBytes(secret)).toHexString().substring(2);
+
+    keystore.pubkey = PublicKey.fromPrivateKey(PrivateKey.fromBytes(secret)).toHexString().replace("0x", "");
     keystore.path = path;
 
     return keystore;
@@ -107,7 +58,7 @@ export class Keystore implements IKeystore {
   public decrypt(password: string): Buffer {
 
     const decryptionKey: bytes = this.kdf(password, this.crypto.kdf.params);
-        
+
     const cipher = AES_128_CTR(decryptionKey.slice(0, 16), this.crypto.cipher.params.iv);
     let decryptedSecret = cipher.update(this.crypto.cipher.message);
     decryptedSecret = Buffer.concat([decryptedSecret, cipher.final()]);
@@ -117,7 +68,7 @@ export class Keystore implements IKeystore {
 }
 
 export class Pbkdf2Keystore extends Keystore {
-  constructor(){
+  constructor() {
     super();
     this.crypto.kdf.function = "pbkdf2";
     this.crypto.kdf.params = DefaultPBKDF2Params;
@@ -127,7 +78,7 @@ export class Pbkdf2Keystore extends Keystore {
 }
 
 export class ScryptKeystore extends Keystore {
-  constructor(){
+  constructor() {
     super();
     this.crypto.kdf.function = "scrypt";
     this.crypto.kdf.params = DefaultScryptParams;
